requirements-optional.txt
# Optional dependencies for advanced features
# Install with: pip install -r requirements-optional.txt

# Token counting (requires Visual Studio Build Tools on Windows)
tiktoken==0.7.0

# Document processing
pypdf2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4

# Embeddings and vector search
numpy==1.24.3
sentence-transformers==2.2.2
faiss-cpu==1.7.4
chromadb==0.4.22

# LangChain integration
langchain==0.1.0
langchain-community==0.0.10
